CMakeLists.txt: Defaulting to shared libraries
In my humble opinion, shared libraries have many advantages, COW being
one of them, the option to update a library without massive rebuilds
is also an advantage.

Using BUILD_SHARED_LIBS and defaulting to ON. This can easily be
overwritten at compile time with `-DBUILD_SHARED_LIBS=OFF`. This is
the common way and embedded build system can configure the library to
be static, if the benefits of a static library might have a higher impact
than using shared libs.

Signed-off-by: Matthias Schoepfer <m.schoepfer@rethinkrobotics.com>