CMakeLists.txt: add proper package for motordriver / master
This should improve the handling of the package in a larger context.
The file_glob() pattern is discouraged, as it may add files to the
build that are not requested or slip into the buid accidentally.

The added lines are along the documentation of [this cmake guide](https://cmake.org/cmake/help/v3.27/guide/importing-exporting/index.html)
It will install the proper cmake package files so that the motordriver lib
can be found outside of this package. Also it can be properly included
in the master subdirectory. Since it is a target now, we no longer need
to set the includes, which was done in an awkward way anyhow.

After installation, other cmake projects can easily find and use
motordriver by
```
find_package(motordriver)
...
target_link_libraries(<yourtargethere> PUBLIC 1x::motordriver)
```

Signed-off-by: Matthias Schoepfer <m.schoepfer@rethinkrobotics.com>